refactor(structured-logs): clean imports visibility and dead helpers
